[WIP] Nexus wall/structure armor changes
Changed the values for Nexus wall/structure defense upgrades.
Gives now one upgrade for Gamma 1 and 2, two for Gamma 3 and 4, and three for Gamma 6.
Changed values for Nexus only upgrades 10 and 11. The change for upgrade 11 seems very high but if my calculations are right it's necessary to keep Nexus' walls/structures strong enough against the Rail Gun in Gamma 9.