Swap cyborg factory throttles in Gamma 1
The western cyborg factory in the SW base has a shorter time to produce its cyborgs than the eastern one. With the result that both groups attack the player at the same time. With an unlucky timing, it's possible that the player has to face the units from 4 factories at the same time. Therefore the factory throttles of the base 2 cyborg factories get swapped.